+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Convert to Template Literal now works on JSX attributes which were not already wrapped with braces. Concretely, `<MyComponent prop="test" />` will now produce `<MyComponent prop={`test`} />` instead of failing.
+- Don't propose Convert to Template Literal on imports
 
 ### Added